requirements.txt
matplotlib==3.3.2
scikit-image==0.17.2
numpy==1.19.2
Pillow==7.2.0

opencv-python~=4.4.0.44